refactor: split index.ts monolith into modules (MEM-130)
Split 1750-line src/index.ts into focused modules:
- src/config.ts — loadConfig, env resolution
- src/api.ts — makeRequest, wallet auth, x402 client
- src/tools.ts — TOOLS array definitions
- src/handlers.ts — all tool handler logic
- src/format.ts — formatMemory, withConcurrency, validators
- src/index.ts — thin server entry point (~50 lines)

Also reads version from package.json instead of hardcoding it.
All 141 tests pass unchanged.
